feature_store/requirements.txt
boto3
sagemaker
pandas
numpy
pyyaml
pyarrow


